Fix data types.
