+++ pysparkling/sql/column.py
@@ -514,7 +514,7 @@ class Column:
 
         if isinstance(dataType, str):
             # pylint: disable=import-outside-toplevel, cyclic-import
-            from pysparkling.sql.ast.ast_to_python import parse_data_type
+            from pysparkling.sql._ast.ast_to_python import parse_data_type
             dataType = parse_data_type(dataType)
         elif not isinstance(dataType, DataType):
             raise NotImplementedError(f"Unknown cast type: {dataType}")
@@ -632,7 +632,7 @@ class Column:
             return self.expr.output_fields(schema)
         return [StructField(
             name=self.col_name,
-            dataType=self.data_type,
+            dataType=self.data_type(schema),
             nullable=self.is_nullable
         )]
