+++ pysparkling/sql/types.py
@@ -676,7 +676,7 @@ class StructType(DataType):
     @classmethod
     def fromDDL(cls, string):
         # pylint: disable=import-outside-toplevel, cyclic-import
-        from .ast.ast_to_python import parse_schema
+        from ._ast.ast_to_python import parse_schema
         return parse_schema(string)
