+++ setup.cfg
@@ -5,7 +5,7 @@ max-line-length = 119
 
 [tool:pytest]
 addopts = --doctest-modules --cov=pysparkling --cov-report html --cov-report xml:reports/coverage.xml --cov-branch
-testpaths = pysparkling
+testpaths = pysparkling tests
 doctest_optionflags = ALLOW_UNICODE NORMALIZE_WHITESPACE
 
 [pycodestyle]
