+++ tests/sql/_ast/test_type_parsing.py
@@ -4,7 +4,7 @@ import io
 import pytest
 
 from pysparkling import Context
-from pysparkling.sql.ast.ast_to_python import parse_data_type, SqlParsingError
+from pysparkling.sql._ast.ast_to_python import parse_data_type, SqlParsingError
 from pysparkling.sql.session import SparkSession
 from pysparkling.sql.types import (
     ArrayType, BinaryType, BooleanType, ByteType, DateType, DecimalType, DoubleType, FloatType, IntegerType, LongType,
