+++ tests/sql/test_session.py
