+++ tests/test_cache.py
