+++ tests/test_stat_counter.py
